server.js refactored, sessions and cookies dependencies installed
