+++ index.js
@@ -1,18 +1,4 @@
-const express = require('express');
-const helmet = require('helmet');
-const cors = require('cors');
-const bcrypt = require('bcryptjs');
-
-const db = require('./data/dbConfig.js');
-const Users = require('./users/users-model.js');
-
-const server = express();
-
-server.use(helmet());
-server.use(express.json());
-server.use(cors());
-
-
+const server = require('./api/server.js');
 
 const port = process.env.PORT || 5000;
 server.listen(port, () => console.log(`\n** Running on port ${port} **\n`));
