+++ package.json
@@ -5,8 +5,10 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "connect-session-knex": "^1.4.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "express-session": "^1.16.2",
     "helmet": "^3.18.0",
     "knex": "^0.17.6",
     "sqlite3": "^4.0.9"
